requirements.txt
# Deep Hedging Research Pipeline
# Core Dependencies

# Deep Learning
torch>=2.0.0
torchvision>=0.15.0

# Scientific Computing
numpy>=1.24.0
scipy>=1.10.0
pandas>=2.0.0

# Finance
yfinance>=0.2.28
nsepy>=0.8
nsetools>=1.0.11

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0

# Optimization & Hyperparameter Tuning
optuna>=3.2.0

# Reinforcement Learning
stable-baselines3>=2.0.0
gymnasium>=0.28.0

# Signature methods (optional - uses pure PyTorch fallback if not available)
# signatory>=1.2.6  # Commented out - requires OpenMP, use PyTorch fallback
esig>=0.9.0

# Experiment Tracking
mlflow>=2.5.0
tensorboard>=2.13.0

# Statistical Testing
statsmodels>=0.14.0
arch>=6.1.0

# Utilities
tqdm>=4.65.0
pyyaml>=6.0
python-dotenv>=1.0.0
joblib>=1.3.0

# Report Generation
jupyter>=1.0.0
nbconvert>=7.6.0

# Testing
pytest>=7.4.0
